Migrate normalize tests
