+++ src/webgpu/shader/execution/expression/call/builtin/normalize.spec.ts
@@ -9,10 +9,10 @@ Returns a unit vector in the same direction as e.
 import { makeTestGroup } from '../../../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../../../gpu_test.js';
 import { TypeF32, TypeVec } from '../../../../../util/conversion.js';
-import { normalizeInterval } from '../../../../../util/f32_interval.js';
+import { FP } from '../../../../../util/floating_point.js';
 import { vectorF32Range } from '../../../../../util/math.js';
 import { makeCaseCache } from '../../case_cache.js';
-import { allInputSources, generateVectorToVectorCases, run } from '../../expression.js';
+import { allInputSources, run } from '../../expression.js';
 
 import { builtin } from './builtin.js';
 
@@ -20,22 +20,46 @@ export const g = makeTestGroup(GPUTest);
 
 export const d = makeCaseCache('normalize', {
   f32_vec2_const: () => {
-    return generateVectorToVectorCases(vectorF32Range(2), 'finite', normalizeInterval);
+    return FP.f32.generateVectorToVectorCases(
+      vectorF32Range(2),
+      'finite',
+      FP.f32.normalizeInterval
+    );
   },
   f32_vec2_non_const: () => {
-    return generateVectorToVectorCases(vectorF32Range(2), 'unfiltered', normalizeInterval);
+    return FP.f32.generateVectorToVectorCases(
+      vectorF32Range(2),
+      'unfiltered',
+      FP.f32.normalizeInterval
+    );
   },
   f32_vec3_const: () => {
-    return generateVectorToVectorCases(vectorF32Range(3), 'finite', normalizeInterval);
+    return FP.f32.generateVectorToVectorCases(
+      vectorF32Range(3),
+      'finite',
+      FP.f32.normalizeInterval
+    );
   },
   f32_vec3_non_const: () => {
-    return generateVectorToVectorCases(vectorF32Range(3), 'unfiltered', normalizeInterval);
+    return FP.f32.generateVectorToVectorCases(
+      vectorF32Range(3),
+      'unfiltered',
+      FP.f32.normalizeInterval
+    );
   },
   f32_vec4_const: () => {
-    return generateVectorToVectorCases(vectorF32Range(4), 'finite', normalizeInterval);
+    return FP.f32.generateVectorToVectorCases(
+      vectorF32Range(4),
+      'finite',
+      FP.f32.normalizeInterval
+    );
   },
   f32_vec4_non_const: () => {
-    return generateVectorToVectorCases(vectorF32Range(4), 'unfiltered', normalizeInterval);
+    return FP.f32.generateVectorToVectorCases(
+      vectorF32Range(4),
+      'unfiltered',
+      FP.f32.normalizeInterval
+    );
   },
 });
