+++ src/webgpu/shader/execution/expression/expression.ts
@@ -29,7 +29,6 @@ import {
   VectorMatrixToVector,
   VectorPairToVector,
   VectorScalarToVector,
-  VectorToVector,
 } from '../../../util/f32_interval.js';
 import { FPInterval } from '../../../util/floating_point.js';
 import {
@@ -814,52 +813,6 @@ function packScalarsToVector(
 export type IntervalFilter =
   | 'finite' // Expected to be finite in the interval numeric space
   | 'unfiltered'; // No expectations
-
-/**
- * @returns a Case for the param and vector of intervals generator provided
- * @param param the param to pass in
- * @param filter what interval filtering to apply
- * @param ops callbacks that implement generating an vector of acceptance
- *            intervals for a vector.
- */
-function makeVectorToVectorCase(
-  param: number[],
-  filter: IntervalFilter,
-  ...ops: VectorToVector[]
-): Case | undefined {
-  param = param.map(quantizeToF32);
-  const param_f32 = param.map(f32);
-
-  const vectors = ops.map(o => o(param));
-  if (filter === 'finite' && vectors.some(v => v.some(e => !e.isFinite()))) {
-    return undefined;
-  }
-  return {
-    input: [new Vector(param_f32)],
-    expected: anyOf(...vectors),
-  };
-}
-
-/**
- * @returns an array of Cases for operations over a range of inputs
- * @param params array of inputs to try
- * @param filter what interval filtering to apply
- * @param ops callbacks that implement generating an vector of acceptance
- *            intervals for a vector.
- */
-export function generateVectorToVectorCases(
-  params: number[][],
-  filter: IntervalFilter,
-  ...ops: VectorToVector[]
-): Case[] {
-  return params.reduce((cases, e) => {
-    const c = makeVectorToVectorCase(e, filter, ...ops);
-    if (c !== undefined) {
-      cases.push(c);
-    }
-    return cases;
-  }, new Array<Case>());
-}
 
 /**
  * @returns a Case for the params and vector of intervals generator provided
