+++ src/webgpu/util/f32_interval.ts
@@ -6,10 +6,6 @@ import { FPInterval, FPMatrix, FPVector, IntervalBounds, FP } from './floating_p
 
 // Interfaces
 
-export interface VectorToVector {
-  (x: number[]): FPVector;
-}
-
 export interface VectorPairToVector {
   (x: number[], y: number[]): FPVector;
 }
@@ -146,10 +142,6 @@ export function multiplicationMatrixVectorInterval(x: number[][], y: number[]): 
 
 export function multiplicationVectorMatrixInterval(x: number[], y: number[][]): FPVector {
   return FP.f32.multiplicationVectorMatrixInterval(x, y);
-}
-
-export function normalizeInterval(n: number[]): FPVector {
-  return FP.f32.normalizeInterval(n);
 }
 
 export function reflectInterval(x: number[], y: number[]): FPVector {
